panvk: Add a central panvk_cmd_draw() helper
This unifies the setup code between direct and indirect draws a bit.
There's now one panvk_cmd_draw() helper which calls into the prepare*()
helpers and then calls a direct or indirect launch_*draw() helper.  This
new helper takes a copy of the pahvk_draw_info, not pointer, because it
may modify it as part of executing the polygon pipeline.

Signed-off-by: Olivia Lee <olivia.lee@collabora.com>